Add ugly hack for detecting unmerged paths
This patch is an ugly hack that fixes a bug in the hint detection for
unmerged paths. It does this by checking the proper header immediately.

The correct solution to this issue is to simply check for all possible
localized values of the section headers from the git localization source
code instead of relying on heuristics about the git commands used in
hints. But this will work to fix my individual case in the mean time.